[CastIt.Android] Clean the house part III
