+++ CastIt.Android/lib/application/main/main_state.dart
@@ -1,0 +1,16 @@
+part of 'main_bloc.dart';
+
+@freezed
+class MainState with _$MainState {
+  factory MainState.loading() = _LoadingState;
+
+  factory MainState.loaded({
+    required String appTitle,
+    required AppThemeType theme,
+    required AppAccentColorType accentColor,
+    required bool initialized,
+    required bool firstInstall,
+    required LanguageModel language,
+    @Default(0) int currentSelectedTab,
+  }) = _LoadedState;
+}
