+++ CastIt.Android/lib/application/played_file_options/played_file_options_state.dart
