+++ CastIt.Android/lib/application/playlist_rename/playlist_rename_event.dart
@@ -1,0 +1,12 @@
+part of 'playlist_rename_bloc.dart';
+
+@freezed
+class PlayListRenameEvent with _$PlayListRenameEvent {
+  factory PlayListRenameEvent.load({
+    required String name,
+  }) = _Load;
+
+  factory PlayListRenameEvent.nameChanged({
+    required String name,
+  }) = _NameChanged;
+}
