+++ CastIt.Android/lib/application/server_ws/server_ws_state.dart
@@ -1,0 +1,13 @@
+part of 'server_ws_bloc.dart';
+
+@freezed
+class ServerWsState with _$ServerWsState {
+  factory ServerWsState.loading() = _LoadingState;
+
+  factory ServerWsState.loaded({
+    required String castItUrl,
+    bool? isConnectedToWs,
+    int? connectionRetries,
+    AppMessageType? msgToShow,
+  }) = _LoadedState;
+}
