--- CastIt.Android/lib/bloc/intro/intro_event.dart
@@ -1,17 +1,0 @@
-part of 'intro_bloc.dart';
-
-@freezed
-class IntroEvent with _$IntroEvent {
-  factory IntroEvent.load() = IntroLoadEvent;
-  factory IntroEvent.changePage({
-    required int newPage,
-  }) = IntroChangePageEvent;
-  factory IntroEvent.urlWasSet({
-    required String url,
-  }) = IntroUrlSetEvent;
-  factory IntroEvent.languageChanged({
-    required AppLanguageType newLang,
-  }) = IntroLanguageChangedEvent;
-
-  const IntroEvent._();
-}
