--- CastIt.Android/lib/bloc/settings/settings_event.dart
@@ -1,30 +1,0 @@
-part of 'settings_bloc.dart';
-
-@freezed
-class SettingsEvent with _$SettingsEvent {
-  factory SettingsEvent.load() = SettingsLoadEvent;
-
-  factory SettingsEvent.connected({
-    required ServerAppSettings settings,
-  }) = SettingsConnectedEvent;
-
-  factory SettingsEvent.disconnected() = SettingsDisconnectedEvent;
-
-  factory SettingsEvent.themeChanged({
-    required AppThemeType theme,
-  }) = SettingsThemeChangedEvent;
-
-  factory SettingsEvent.castItUrlChanged({
-    required String castItUrl,
-  }) = _SettingsCastItUrlChangedEvent;
-
-  factory SettingsEvent.accentColorChanged({
-    required AppAccentColorType accentColor,
-  }) = SettingsAccentColorChangedEvent;
-
-  factory SettingsEvent.languageChanged({
-    required AppLanguageType lang,
-  }) = SettingsLanguageChangedEvent;
-
-  const SettingsEvent._();
-}
